# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World Pool API",
        "description": "API para gerenciamento de bolões da Copa do Mundo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
        "/api/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Lista todos os grupos da copa",
                "operationId": "f286cbf34dc4be32a94459a335658168",
                "responses": {
                    "200": {
                        "description": "Lista de grupos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/groups/{id}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Retorna detalhes de um grupo específico",
                "operationId": "38a1f4990d0dc69ead7f7377ad016117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grupo encontrado"
                    },
                    "404": {
                        "description": "Grupo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/pools/{poolId}/leaderboard": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Retorna o ranking completo do bolão",
                "operationId": "49dfcbdcec56bf3bada4a815fbd80928",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking do bolão"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/top": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Retorna os primeiros colocados do ranking",
                "operationId": "baf5fbd316279ea7b194ff120e54a7b9",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top colocados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/me": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Retorna a posição do usuário autenticado no ranking",
                "operationId": "3c3d14911a1cb5ac8c86e35d9f62e23f",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posição do usuário no ranking"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/{userId}": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Retorna as estatísticas de ranking de um participante específico",
                "operationId": "14f742d015c6a5527c253581fa3c7bd8",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estatísticas do participante"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Remove o usuário do ranking do bolão (interno)",
                "operationId": "c35ccdecbd7fff08c5423199bf309d51",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido do ranking"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/recalculate": {
            "post": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Recalcula o ranking após finalização de partidas",
                "operationId": "96dbfb61abe4cb42aac490fdded8bff4",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking recalculado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/{userId}/sync": {
            "post": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Atualiza o ranking de um usuário específico no bolão (interno)",
                "operationId": "86e7803e8458b5be65c82f5984a8c10f",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking do usuário sincronizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/{userId}/create": {
            "post": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Cria entrada de ranking para um usuário ao entrar no bolão (interno)",
                "operationId": "195f6996428b9231ef714f3de16f71b5",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Entrada criada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pools/{poolId}/leaderboard/rebuild": {
            "post": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Reconstrói completamente o leaderboard com base nos palpites existentes (interno)",
                "operationId": "e4db2cb4adf4cd422c785f0f21e5ff8d",
                "parameters": [
                    {
                        "name": "poolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leaderboard reconstruído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token obtido via POST /api/login",
                "scheme": "bearer"
            }
        }
    },
...
}